Cleanup future codebase
