+++ Sources/Core/Streams/Base64Stream.swift
