+++ Sources/Core/Streams/Stream.swift
